añadido stepper para venta

+++ src/main/java/com/hbourgeot/inversiones7h/controllers/CompraController.java
@@ -123,7 +123,7 @@ public class CompraController implements BootInitializable {
     
     compra.setProducto(producto);
     compra.setProveedor(proveedor);
-    compra.setCantidad(cantidad);
+    compra.setCantidad(cantidad.longValue());
     compra.setMontoTotal(montoTotal);
 
     // Guardar la compra utilizando el servicio CompraService
